+++ __mocks__/ssr-window.js
@@ -1,0 +1,7 @@
+module.exports = {
+  extend: jest.fn(),
+  getDocument: jest.fn(),
+  getWindow: jest.fn(),
+  ssrDocument: {},
+  ssrWindow: {},
+};