+++ libraries/common/components/Swiper/components/SwiperItem/styles.js
@@ -3,4 +3,4 @@ import { css } from 'glamor';
 export const item = css({
   position: 'relative',
   height: '100%',
-});
+}).toString();