+++ libraries/common/components/Swiper/styles.js
@@ -12,6 +12,7 @@ export const container = css({
 
 export const innerContainer = css({
   overflow: 'hidden',
+  '--swiper-navigation-color': themeColors.gray,
   ' .swiper-wrapper': {
     alignItems: 'stretch',
   },
@@ -19,7 +20,6 @@ export const innerContainer = css({
     height: 'auto',
   },
   ' .swiper-pagination': {
-    transform: 'translate3d(-50%,0,0) !important',
     ' .swiper-pagination-bullet': {
       background: themeColors.gray,
       opacity: '.5',
@@ -47,8 +47,16 @@ export const wrapper = css({
 
 export const buttonNext = css({
   backgroundImage: "url(\"data:image/svg+xml;charset=utf-8,%3Csvg%20xmlns%3D'http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg'%20viewBox%3D'0%200%2027%2044'%3E%3Cpath%20d%3D'M27%2C22L27%2C22L5%2C44l-2.1-2.1L22.8%2C22L2.9%2C2.1L5%2C0L27%2C22L27%2C22z'%20fill%3D'%23808080'%2F%3E%3C%2Fsvg%3E\") !important",
+  // Hide original SwiperJS chevron
+  ':after': {
+    color: 'transparent',
+  },
 }).toString();
 
 export const buttonPrev = css({
   backgroundImage: "url(\"data:image/svg+xml;charset=utf-8,%3Csvg%20xmlns%3D'http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg'%20viewBox%3D'0%200%2027%2044'%3E%3Cpath%20d%3D'M0%2C22L22%2C0l2.1%2C2.1L4.2%2C22l19.9%2C19.9L22%2C44L0%2C22L0%2C22L0%2C22z'%20fill%3D'%23808080'%2F%3E%3C%2Fsvg%3E\") !important",
+  // Hide original SwiperJS chevron
+  ':after': {
+    color: 'transparent',
+  },
 }).toString();