+++ themes/theme-gmd/widgets/ProductSlider/index.jsx
@@ -136,17 +136,20 @@ class ProductSlider extends PureComponent {
         {this.renderHeadline()}
         <ProductListTypeProvider type="productSlider" subType="widgets">
           <Swiper
-            autoPlay={sliderSettings.autostart}
+            {...sliderSettings.autostart && {
+              autoplay: {
+                delay: Number.parseInt(sliderSettings.delay, 10),
+              },
+            }}
             loop={false}
             indicators={false}
             controls={false}
-            interval={Number.parseInt(sliderSettings.delay, 10)}
             freeMode
             slidesPerView={slidesPerView}
             classNames={{ container: styles.sliderContainer }}
           >
             {products.slice(0, PRODUCT_SLIDER_WIDGET_LIMIT).map(product => (
-              <Swiper.Item key={product.id} className={styles.sliderItem}>
+              <Swiper.Item key={product.id}>
                 <ProductListEntryProvider productId={product.id}>
                   <Card className={styles.card}>
                     <ProductCard